Fix compiler issues on MacBook
There is some compilation errors on some MacBook when the cpu type doesn't match with the kernel version. eg. if the cpu is i386 but the kernel is 64bit (x86_64), the build will be i386 and terraform will not work.

Fix the terraform error: 
Error: Failed to instantiate provider "cloudamqp" to obtain schema: fork/exec /Users/username/.terraform.d/plugins/terraform-provider-cloudamqp: bad CPU type in executable