Add a link to the API location (#73)
So I can determine the resource_id to import, I've added a link to the API docs.

I also corrected a small typo on the import template structure. I believe `rmq_url` is not an intention revealing name to be used to name an instance (it sounds like a param) but I'll leave that to someone else to update if they want.