Enable faster instance destroy options in provider configuration (#209)
When tearing down the resource block, the cloudamqp_instance and underlying servers will be deleted. If resources such as plugins also are present, they will use their delete behaviour (disable plugins in this case). This action is not necessary since the servers are deleted.

Adds configurable global options in the provider configuration called 'enable_faster_instance_destroy' to skip the delete behaviour for.

- cloudamqp_plugin
- cloudamqp_plugin_community
- cloudamqp_security_firewall