feature(terraform-sdk): Use tf-sdk-migrator to migrate to sdk version 2
Most of the heavy lifting is done by the tool and I just had to change the
signature of some lambdas to include the newly added context.

Som deprecations have been added as a result of this but we can be dealt with
it other MR.

This drops support for terraform <= 0.11.