Use placeholder string for Cloudwatch Key ID
This allows the recording to use a real value (because the backend does
some verification of it) but these tests only need to care about the
placeholder value.

We also doesn't need to know how the response look in order to sanitize
it. This approach should be used for most of the "blocked fields".