Enable import of VPC peering resource for AWS (#309)
### WHY are these changes introduced?

The resource `cloudamqp_vpc_peering` lacks import functionality. This got requested way back in: https://github.com/cloudamqp/terraform-provider-cloudamqp/issues/158.
We found a way to enable this for GCP https://github.com/cloudamqp/terraform-provider-cloudamqp/pull/308, take a similar approach for AWS. 

### WHAT is this pull request doing?

- Adds import functionality to the `cloudamqp_vpc_peering` resource
- Updates documentation
- Adds Go VCR test for accept and import

### HOW can this pull request be tested?

Run VCR test for accepting the peering and importing the resource.
Limited from the complete flow to not mix in the AWS provider in the test.