Migrate cloudamqp_rabbitmq_configuration towards Terraform plugin framework (#346)
### WHY are these changes introduced?

Part of Terraform plugin framework migration #281.

Older Terraform plugin SDK versions cannot detect if an argument
have been set or not in the configuration file. If not set, the default value is
used for resp. data type (int will be automatically assigned 0).

Known issue for this resource, to not be able to set heartbeat to 0. Reported in #214 

### WHAT is this pull request doing?

- Adds API model for request/response for the RabbitMQ configuration
- Migrates `cloudamqp_rabbitmq_configuration` resource
- Adds new VCR-test to test zero values
- Updates docs